Adding support for opening new browser windows in .NET
Since support has been added to the W3C WebDriver Specification for
creating a new top-level browser context (a "new tab" or "new window"),
this change implements the new command in the .NET language bindings.
The method creates a new browser tab/window and switches to it. As
an argument, the method takes a type hint to indicate what type of
browser context the user wants (a tab or a window). Example usage:

    driver.SwitchTo().NewWindow(WindowType.Tab);

Note carefully, however, this type hint is only a suggestion. If a
browser driver does not support creating new top-level browsers of
the type desired by the user, it will still create the new top-level
browser using whatever type it does support. This is perfectly
compliant with the specification language for this command. As a
concrete example, at the time of this commit, executing the above
sample code against the IE driver will create a new top-level browser
context, but as a window, not a tab, in spite of the fact that a new
tab was explicitly requested. Even though the browser itself supports
creation of new tabs, the driver does not. This will not be considered
a valid bug, and issue reports claiming it is so will be summarily
closed, possibly without comment.