Disabling non-spec-compliant tests in .NET test suite
Most WebDriver implementations are now compliant with the W3C WebDriver
Specification. As such, they do not respond to end points that predate
the spec, like the logging end points, or those specific to HTML5
constructs like LocalStorage or SessionStorage. Rather than ignore the
tests for every implementation, we are removing them entirely for now.