Updating .NET bindings to be based on System.Net.Http.HttpClient
With the advent of .NET Core 2.0, the preferred mechanism for issuing
HTTP requests is with the use of System.Net.Http.HttpClient. The .NET
bindings could not use that class before because it was not supported
on earlier versions of the .NET Framework. Now that support for those
earlier versions has been removed, the bindings can take advantage of
the more efficient communication mechanism.