Updating .NET ExecutingAsyncJavascriptTest to create new driver after test
When using the OSS protocol dialect, ChromeDriver does not handle unexpected
alerts encountered when executing async JavaScript. If the alerts tests
are enabled, many remaining tests fail because the alert is left active.
The tests are still enabled, because when ChromeDriver uses the W3C protocol
dialect, the alerts are properly handled. The addition of the attribute on
the test class is to ensure that when running with the legacy OSS protocol
dialect, that those cascading failures are eliminated. The attribute can
be removed once ChromeDriver fully supports the W3C dialect and has made that
dialect the default.s