[grid] Ignoring NoSuchSessionException in SESSION_CLOSED event listener on a node.
This exception is thrown if the session with the given id does not exist on this node. When a node stops the session on driver.quit() call it notifies all the parties by sending SESSION_CLOSED event to MQ. The distributor handles this event and commands the "local node" to stop the session. But if the distributor and the node are working in the same process the session is already closed on the "local node" and it throws. Ignoring this exception looks safe because we can't do (and shouldn't do) anything in this case.