Adding .NET integration tests for setting a proxy
This commit now brings the .NET test suite into alignment with the Java
test suite for this feature. It includes a forked version of FryProxy
modified to remove its dependency on log4net, and a few other minor
modifications. The forked changes can be found at
https://github.com/jimevans/FryProxy. The resulting NuGet package has
been added to the local package repository, but is not available from
NuGet directly.