Use W3C version of Edge (version 18)
This commit updates implementation of Edge browser via MicrosoftWebDriver
to the recent stable version which uses W3C mode by default.
This renders unnecessary a custom bridge implementation which mixed
OSS and W3C dialects. The new version is tested on Edge version 18,
though we might need to change it when a new Chromium-based version
will be released.