Start providing abstractions for consistent handling of HttpRequests
The new grid server makes an effort not to differentiate between
local and remote execution. This is a nice model, and we should
extend it to other places in the code base.

This also provides an abstraction to start solving the problem of
wanting to wrap an HttpClient with methods that modify requests
or responses (eg. to add authentication headers).

Finally, we provide enough infrastructure to make providing
routes to various handlers easy to implement, allowing complicated
routing mechanisms to be built up.