[js] "fix" actions tests for chrome by wrapping spec in an extra describe block.
This appears to be necessary for the afterEach hook to properly
run and clear action state between tests. This appears to be a timing
issue around mocha and not a problem with the driver itself: adding
an extra `driver.actions().clear()` call directly to the test seemed
to work.

I should dig into this more, but this test gets things passing without
unnecessarily slowing things down (to launch a new browser session)
or introducing too big of a hack.