[CDP] Provide an in-process proxy for stubbing responses
This makes use of the CDP to provide a lightweight mechanism
for intercepting requests from the browser to the server,
allowing them to be overridden if necessary.

In order to do this, the user needs to be provide a `Route`
from our own HTTP abstractions. If the route matches, then
it is called to figure out what to return to the browser.

This allows tests to stub out backends if they see this as
being necessary.