[bazel] Use Java 11 as the base JDK
Selenium continues to target Java 8, but some functionality has to use
Java 11 in order to work properly. Until we have the equivalent of
Animal Sniffer, we'll trust to luck and good fortune that we won't
accidentally use Java 11 APIs in Java 8 classes.

This change is needed in order to allow the `java_export` rule to
generate java module-info classes automagically.