Updating .NET build to use new Bazel build rules
The rules used for building the .NET binaries can now be found
at https://github.com/Brightspace/rules_csharp. These are not the
"official" Bazel build rules for .NET projects, but they are more
flexible and better suited to the needs of the Selenium project.
Note that they are still somewhat under active development, and
it may be some time before `bazel test` will work with the .NET
portion of Selenium. In the meantime, users can manually use the
NUnit console checked into the `third_party` directory to run
the tests in the .NET test suite.