Adding support for relative locators for .NET
To use relative locators in .NET, take advantage of the RelativeBy
class to use with FindElement and FindElements. This change also
marks the IFindsBy* interfaces (IFindsById, IFindsByName, etc.) as
deprecated. The presence of individual methods for finding elements
by each mechanism is an artifact of an older architecture, and with
the standardization of the locator methods in the W3C WebDriver
Specification, they are now just excess code, and will be deleted.