[java] Fail faster if a driver server can't start.
A user can specify path to a driver server using system properties. If the target file is not valid (corrupted or compiled for other operating system/architecture) the process will die prematurely, there is no need to poll for server status. So we have to wait for any of three conditions: 1) process finished, 2) status is OK, 3) timeout.