Updating .NET common tests to work for multiple framework versions
The .NET tests now support being run under both the .NET Framework
and .NET Core. This also now allows the user to run the tests using
Bazel, and can run against any of the supported browsers. By default,
the tests run using .NET Core. To run the tests at the command line,
use the following command:

    bazel test //dotnet/test/common:<browser> --enable_runfiles

where <browser> is one of the following values:

 * chrome
 * chromedev
 * firefox
 * firefoxnightly
 * ie
 * edge
 * edgelegacy
 * safari
 * safaritechpreview

To specify to run the driver using the .NET Framework, append:

    --define framework=netfx

to the command line.

To specify where the browser driver (chromedriver, geckodriver,
IEDriverServer, etc.) can be found, append the following to the
command line:

    --test_arg=--params=DriverServiceLocation=<path>

where <path> is the path to the executable.

The command line rule is only verified to work on Windows at the
moment. There is an attempt at support for non-Windows OSes, but
that has not been tested, and may need to be fixed in a future
commit.