case insensitive check for the content-length header
According to RFC 2616 and 7230, headers names are case insensitive.

When creating a new session, the webdriver server is checking the
request for a case sensitive Content-Type header. This causes the body
to be empty on session creation and will result in the IEDriverServer
complaining about missing capabilities.

This change will search for the Content-Type in a case insensitive
manner before capturing the value of the header.

Fixes #7986

Signed-off-by: Jim Evans <james.h.evans.jr@gmail.com>