Cleanup the .bazelrc (#8281)
Note that all build options are inherited to the test command, and that all
test-specific options are also available on the build command. By using
"build" instead of "test", we ensure that the configuration options are the
same so that Bazel can cache the analysis (and execution) phases (otherwise
it can't and switching back and forth can be very slow).