[cdp] Implement an idealized CDP protocol over v84 of CDP
We isolate Selenium from the fact that the CDP changes frequently by
presenting our own API to use (the `idealized` package) and
implementing that separately for each version of the protocol we
use. It's not elegant, but it works.

This change modifies that idealized CDP and makes it work with v84 of
the existing CDP, proving out that the changes work.