Introduce version-independent implementation of CDP for .NET
This commit introduces a facade around Chromium's Developer Tools
Protocol, which will present developers with a stable,
version-independent API for the use cases most relevant to Selenium. It
is still possible for users to have access to the version-specific API
by calling the `GetVersionSpecificDomains<T>` method on the DevTools
session object created when the user calls
`IDevTools::CreateDevToolsSession`. Note that using this method will tie
any consuming code to one and only one version of a Chromium-based
browser, and the consuming code may need source-code changes and
recompilation when new versions of the browser are released.

Additionally, this commit enables dynamic generation of the .NET API for
the DevTools Protocol during the build process.

At this time, only local execution of CDP code is supported. Remote
proxying of CDP calls through the Selenium Grid is planned, but not yet
implemented. Also, certain helper methods to enable common desired
scenarios for Selenium users that may be present in other language
bindings still need to be created for .NET.