[grid] Allowing a path on the host & container to be configurable [skip ci]
When we run inside a container, we need to know where to save the
assets inside it, which is not the same path we have on the host.
The user will mount the host path to the container path, so for
the user the assets will end up on the same path on their host
machine.