[grid] Giving precedence to file configuration over flags
In general, Grids should be configured via files, and before
this change, flags and default values were taking precedence
over files, which broke the expected behaviour. For example,
"relax-checks" in a file configuration was always ignored
because the default value is false.

This fixes https://github.com/SeleniumHQ/docker-selenium/issues/1145