[bazel] Create test suites as sharded tests
We used to create a new `java_test` target per java test source we
found. We now create a single `java_test`, but set the shard count to
be equal to the number of tests.

This will create fewer targets in the build graph, and should lead to
faster build times since fewer files are being created. Test run time
_should_ be unaffected, since bazel's junit4 runner is aware of test
sharding and appears to know how to break a test suite up so that it
can run efficiently when sharded.