[grid] Simplifying DefaultSlotSelector logic
The idea has always been to keep as available
as possible the Nodes with diverse configurations.
Such as the ones supporting different browsers,
like a Node supporting Chrome, Firefox and Safari.

But if we have a Node that only supports Chrome,
we'd like to give as many Chrome sessions as possible
to keep the one with Safari as available as possible,
so when a Safari session request comes in, it can be
served right away.

The previous logic had that in mind, but the
implementation was rather complex. This simplified
version orders Nodes by the number of browsers they
support. Therefore, it would offer "specialized"
Nodes first (e.g. only supporting one browser), and
the more diverse ones at the bottom.