[grid] Enabling driver configurations to Nodes
Only useful for Firefox, Chrome and Edge Chromium
for now. This allows users to configure a stereotype
that serves as an input for the Slot configuration
in a Node. A user will be able to specify a custom
browser binary location, in case they need to run
tests with beta, dev or nightly versions and so forth.
It is possible to set specific configurations for each
browser inside their prefixed block, like a given log
level for Firefox/GeckoDriver through "moz:firefoxOptions"