[grid] Dynamic Grid working properly on Linux
Docker Networking is a little different in
Linux and the approach of mapping a port and
relying on the Docker daemon url is not enough,
we need to use the container's IP and there is
no need to map a port to the host for that.

And after testing this approach in different OS,
we can make this behaviour the default for
Dynamic Grid, so we do not need to use the PortProber
when we run inside Docker.