[grid] Allowing users to overwrite recommended max sessions
Until now, we have had the limit per Node to run up to
(# of available processors) concurrent sessions on each
host, as the idea is to embrace stability and reliability
to the running browser sessions, as one browser session
should use one processor. Except IE and Safari, which
only run one concurrent session per host.

If a user sets `override-max-sessions` to `true`, the
value for `max-sessions` can be higher than the recommended
max sessions. A set of warnings are printed on the command
line when this behaviour is detected.

Fixes #9300