Move logic for creating a session into the `LocalDistributor`
We originally lifted this logic up to the `Distributor` class when
working on a version of the Distributor that stored state in either a
database or redis. However, we never implemented those, and there is
restructuring work to do. Moving the logic back into the main
implementation makes this easier.