Rework how the Distributor creates sessions
This brings the distributor more inline with the w3c
spec. Specifically, all the capabilities in the payload will be
examined in order, and a session started for each until one session
actually starts. If the Grid claims to support at least one payload
but no slots are available, then the request will be retried at some
other time. If no slots support any of the capabilities in the new
session request, then the user will be notified
immediately. Similarly, if every attempt to start a session leads to
an exception, the user will be notified immediately.

There's an argument that we should be more generous in how we
interpret payloads, but for now the approach above seems reasonable.