Implement Ruby gem building with Bazel
This adds 2 new targets to Ruby Bazel build. Each target copies all
necessary files for gem it needs to build and then packages into a gem
file that is stored in Bazel output directory. For now there is no
way to publish gem to RubyGems, but this should be easy to implement.
Please see the following table to understand what commands to use and
what gem files each command generates.

| command                             | output                              |
| bazel build //rb:selenium-devtools  | bazel-bin/rb/selenium-devtools.gem  |
| bazel build //rb:selenium-webdriver | bazel-bin/rb/selenium-webdriver.gem |