Ensure that Capabilities implement `hashCode` and `equals` consistently
The javadocs for hashcode state that:

> If two objects are equal according to the equals(Object) method,
> then calling the hashCode method on each of the two objects must
> produce the same integer result.

Prior to this PR, this was not the case. This was problematic since
the property of hashcodes being equal is used in Collections to
quickly allow them to determine whether the contents are equal.

Where possible, we precompute the hashcode and use that value.