[dotnet] Rename TerminateDevToolsSession to ResetDevToolsSession
For the case where a CDP session is made invalid, for example, because a
window is closed, this method will now reset the CDP session so that the
user can continue to use it without closing and establishing a new web
socket connection.

Also added disposal of DevTools sessions to Firefox and Remote when
required.