[grid] Choosing the driver builder with highest score
Previously we were using all builders that had a score
higher than zero for each set of stereotypes. But this
can lead to erroneous behaviour because two different
builders should not serve the same stereotypes, it will
be misleading for the user.

With this, we only use the builder that has the highest
score. The current behaviour is not affected, and this
paves the way to have insert a DriverService for
Safari Tech Preview.