[grid] Adding SafariTechPreviewDriverService
This enables users to run Safari Tech Preview in
Grid. It was not possible before because the builder
assigned to the stereotype did not have the right
browser driver binary path configured, and setting
that afterwards was not simple.

Also, a couple of things got deprecated in
SafariDriverService, so we can remove them in the
next release.

Fixes #6791