[grid] Rejecting events when a Secret cannot be parsed.
This might happen if Grid is deployed to an environment
like Kubernetes, where checks are usually done on every
open port. This leads to unexpected payloads to be sent
to the EventBus ports, which then cannot be properly
parsed and then the listeners threads are interrupted.

This change tries to parse the Secret and if not possible,
the message is rejected. Also, the thread won't be
interrupted anymore.

Complements solution pushed for #9959