[grid] Using non-loopback address when default config is used
The `--hub` flag was introduced just before the 4.0.0 release,
and this implicitly made the use of `0.0.0.0` as the default
IP for the Hub and hence the Grid.

When someone tried to connect to a remote Grid, through the
`RemoteWebDriver.builder()`, an error was happening since
the `Augmenter` makes the `RemoteWebDriver` connect to
`se:cdp`. Given that the Grid url was `http://0.0.0.0:4444`,
the client was trying to connect to `ws://0.0.0.0:4444/session/...`,
which caused a failure.

This fixes that by retrieving the non-loopback address of the machine,
under the assumption that Hub and Nodes are on the same machine. If the
Hub and Nodes are on different machines, or inside Docker containers,
the users will need to configure the Grid url themselves
(through `--grid-url`).

Fixes #9976