[grid] Adding a "bind-host" flag
#9981 fixed a regression where the Grid
server was binding to all interfaces, even
if a host was being configured.

However, this was not ideal for environments
like Docker, where the host value can be the
reachable IP of the machine where the containers
are running, but we cannot bind to that IP inside
the container.

This adds a flag where the user can decide if the
server should bind to the host or not, true by
default to keep the current behaviour, but it will
be configured to false by default in the docker-selenium
project.

This is a partial fix for https://github.com/SeleniumHQ/docker-selenium/issues/1453