[grid] Improving Node health check execution
This commit does a couple of things...

First, it removes the Regularly used to run all health
checks and now uses a ScheduledExecutorService. This
was done because this replacement has shown better
CPU performance (checked with VisualVM(.

Second, it runs the health check when the Node is
registered to set the availability to UP. This was
already being done by Regularly. But if the check
was failing for any reason, no retry was attempted,
and the next check execution was when the Regularly
schedule interval kicks in. Instead of that, we use
now Failsafe to run a retry when the initial execution
of the health check fails. This is useful in
environments like Kubernetes, since the pod comes up,
but it might take a few seconds until the Distributor
can reach the IP assigned to the pod.

#9847